Update dependencies for ES 7.8.1
