+++ gradle.properties
@@ -1,3 +1,7 @@
-ltrVersion = 1.3.0
-elasticsearchVersion = 7.8.0
+ltrVersion = 1.3.1
+
+# Dependencies versions
+elasticsearchVersion = 7.8.1
 luceneVersion = 8.5.1
+antlrVersion=4.5.1-1
+ow2Version=7.2
