+++ src/main/java/com/o19s/es/ltr/rest/RestSearchStoreElements.java
@@ -1,0 +1,2 @@
+package com.o19s.es.ltr.rest;public class RestSearchStoreElements {
+}
